feat: hide test app
