+++ src/Ivy.Samples.Shared/Apps/Widgets/Inputs/TextInputApp.cs
@@ -11,7 +11,6 @@ public class TextInputApp : SampleBase
                    new Tab("Variants", new TextInputVariants()),
                    new Tab("Sizes", new TextInputSizes()),
                    new Tab("Affixes", new TextInputAffixes()),
-                   new Tab("All Input Affixes", new InputAffixesGallery()),
                    new Tab("Data Binding", new TextInputDataBinding()),
                    new Tab("Length Constraints", new TextInputLengthConstraints()),
                    new Tab("Events", new TextInputEventsTab())
@@ -277,117 +276,4 @@ public class TextInputAffixes : ViewBase
                | nullableState.ToTextInput().Suffix("%").Placeholder("Percentage")
                | nullableState.ToTextInput().Prefix("https://").Suffix(".com").Placeholder("domain");
     }
-}
-
-public class InputAffixesGallery : ViewBase
-{
-    public override object Build()
-    {
-        var textState = UseState("ivy.app");
-        var passwordState = UseState("secret");
-        var searchState = UseState("ivy.app");
-        var emailState = UseState("user@ivy.app");
-        var telState = UseState("5550100");
-        var urlState = UseState("ivy.app");
-        var textareaState = UseState("Notes");
-        var numberState = UseState(42.5m);
-        var currencyState = UseState(Currency.USD);
-        var dateState = UseState(DateTime.Now);
-        var rangeState = UseState((DateOnly.FromDateTime(DateTime.Today), DateOnly.FromDateTime(DateTime.Today.AddDays(7))));
-        var boolState = UseState(true);
-        var colorState = UseState("#6366f1");
-        var feedbackState = UseState(3);
-        var iconState = UseState(Icons.Heart);
-        var codeState = UseState("console.log('ivy');");
-        var shortcutTextState = UseState<string?>((string?)null);
-        var shortcutPasswordState = UseState<string?>((string?)null);
-        var shortcutSearchState = UseState<string?>((string?)null);
-        var shortcutEmailState = UseState<string?>((string?)null);
-        var shortcutTelState = UseState<string?>((string?)null);
-        var shortcutUrlState = UseState<string?>((string?)null);
-        var shortcutTextareaState = UseState<string?>((string?)null);
-        var currencyOptions = typeof(Currency).ToOptions();
-
-        return Layout.Vertical()
-               | Callout.Info(
-                   "Compare prefix, suffix, and both affixes across inputs with transparent affix chrome. Each row is its own four-column grid so columns stay aligned. Shortcut-key rows use empty nullable fields so the kbd hint is visible; with a suffix affix it sits between the field and the suffix.")
-               | Text.H2("Text inputs")
-               | AffixHeaderRow()
-               | AffixRow("Text", textState.ToTextInput().Prefix(Icons.Link), textState.ToTextInput().Suffix(Icons.Globe), textState.ToTextInput().Prefix(Icons.Link).Suffix(Icons.Globe))
-               | AffixRow("Password", passwordState.ToPasswordInput().Prefix(Icons.Lock), passwordState.ToPasswordInput().Suffix(Icons.Key), passwordState.ToPasswordInput().Prefix(Icons.Lock).Suffix(Icons.Shield))
-               | AffixRow("Search", searchState.ToSearchInput().Prefix(Icons.ListFilterPlus).Placeholder("Search..."), searchState.ToSearchInput().Suffix(Icons.Tag).Placeholder("Search..."), searchState.ToSearchInput().Prefix(Icons.Folder).Suffix(Icons.Globe).Placeholder("Search..."))
-               | AffixRow("Email", emailState.ToEmailInput().Prefix(Icons.Mail), emailState.ToEmailInput().Suffix(Icons.AtSign), emailState.ToEmailInput().Prefix(Icons.Mail).Suffix(Icons.AtSign))
-               | AffixRow("Tel", telState.ToTelInput().Prefix(Icons.Phone), telState.ToTelInput().Suffix(Icons.Hash), telState.ToTelInput().Prefix(Icons.Phone).Suffix(Icons.Hash))
-               | AffixRow("Url", urlState.ToUrlInput().Prefix(Icons.Link), urlState.ToUrlInput().Suffix(Icons.ExternalLink), urlState.ToUrlInput().Prefix(Icons.Link).Suffix(Icons.ExternalLink))
-               | AffixRow("Textarea", textareaState.ToTextareaInput().Prefix(Icons.FileText), textareaState.ToTextareaInput().Suffix(Icons.Type), textareaState.ToTextareaInput().Prefix(Icons.FileText).Suffix(Icons.Type))
-               | Text.H2("Shortcut keys with affixes")
-               | AffixHeaderRow()
-               | AffixRow(
-                   "Text",
-                   shortcutTextState.ToTextInput().Nullable().Prefix(Icons.Link).ShortcutKey("Ctrl+/").Placeholder("Path"),
-                   shortcutTextState.ToTextInput().Nullable().Suffix(Icons.Globe).ShortcutKey("Ctrl+/").Placeholder("Domain"),
-                   shortcutTextState.ToTextInput().Nullable().Prefix(Icons.Link).Suffix(Icons.Globe).ShortcutKey("Ctrl+/").Placeholder("URL"))
-               | AffixRow(
-                   "Password",
-                   shortcutPasswordState.ToPasswordInput().Nullable().Prefix(Icons.Lock).ShortcutKey("Ctrl+L").Placeholder("Password"),
-                   shortcutPasswordState.ToPasswordInput().Nullable().Suffix(Icons.Key).ShortcutKey("Ctrl+L").Placeholder("Password"),
-                   shortcutPasswordState.ToPasswordInput().Nullable().Prefix(Icons.Lock).Suffix(Icons.Shield).ShortcutKey("Ctrl+L").Placeholder("Password"))
-               | AffixRow(
-                   "Search",
-                   shortcutSearchState.ToSearchInput().Nullable().Prefix(Icons.ListFilterPlus).ShortcutKey("Ctrl+K").Placeholder("Search..."),
-                   shortcutSearchState.ToSearchInput().Nullable().Suffix(Icons.Tag).ShortcutKey("Ctrl+K").Placeholder("Search..."),
-                   shortcutSearchState.ToSearchInput().Nullable().Prefix(Icons.Folder).Suffix(Icons.Globe).ShortcutKey("Ctrl+K").Placeholder("Search..."))
-               | AffixRow(
-                   "Email",
-                   shortcutEmailState.ToEmailInput().Nullable().Prefix(Icons.Mail).ShortcutKey("Ctrl+E").Placeholder("Email"),
-                   shortcutEmailState.ToEmailInput().Nullable().Suffix(Icons.AtSign).ShortcutKey("Ctrl+E").Placeholder("Email"),
-                   shortcutEmailState.ToEmailInput().Nullable().Prefix(Icons.Mail).Suffix(Icons.AtSign).ShortcutKey("Ctrl+E").Placeholder("Email"))
-               | AffixRow(
-                   "Tel",
-                   shortcutTelState.ToTelInput().Nullable().Prefix(Icons.Phone).ShortcutKey("Ctrl+J").Placeholder("Phone"),
-                   shortcutTelState.ToTelInput().Nullable().Suffix(Icons.Hash).ShortcutKey("Ctrl+J").Placeholder("Extension"),
-                   shortcutTelState.ToTelInput().Nullable().Prefix(Icons.Phone).Suffix(Icons.Hash).ShortcutKey("Ctrl+J").Placeholder("Phone"))
-               | AffixRow(
-                   "Url",
-                   shortcutUrlState.ToUrlInput().Nullable().Prefix(Icons.Link).ShortcutKey("Ctrl+U").Placeholder("URL"),
-                   shortcutUrlState.ToUrlInput().Nullable().Suffix(Icons.ExternalLink).ShortcutKey("Ctrl+U").Placeholder("URL"),
-                   shortcutUrlState.ToUrlInput().Nullable().Prefix(Icons.Link).Suffix(Icons.ExternalLink).ShortcutKey("Ctrl+U").Placeholder("URL"))
-               | AffixRow(
-                   "Textarea",
-                   shortcutTextareaState.ToTextareaInput().Nullable().Prefix(Icons.FileText).ShortcutKey("Ctrl+T").Placeholder("Notes"),
-                   shortcutTextareaState.ToTextareaInput().Nullable().Suffix(Icons.Type).ShortcutKey("Ctrl+T").Placeholder("Notes"),
-                   shortcutTextareaState.ToTextareaInput().Nullable().Prefix(Icons.FileText).Suffix(Icons.Type).ShortcutKey("Ctrl+T").Placeholder("Notes"))
-               | Text.H2("Other inputs")
-               | AffixHeaderRow()
-               | AffixRow("Number", numberState.ToNumberInput().Prefix(Icons.DollarSign).Precision(1), numberState.ToNumberInput().Suffix(Icons.Percent).Precision(1), numberState.ToNumberInput().Prefix(Icons.DollarSign).Suffix(Icons.Coins).Precision(1))
-               | AffixRow("Select", currencyState.ToSelectInput(currencyOptions).Prefix(Icons.DollarSign), currencyState.ToSelectInput(currencyOptions).Suffix(Icons.BadgeDollarSign), currencyState.ToSelectInput(currencyOptions).Prefix(Icons.DollarSign).Suffix(Icons.BadgeDollarSign))
-               | AffixRow("DateTime", dateState.ToDateTimeInput().Prefix(Icons.Calendar), dateState.ToDateTimeInput().Suffix(Icons.Clock), dateState.ToDateTimeInput().Prefix(Icons.Calendar).Suffix(Icons.Clock))
-               | AffixRow("Date range", rangeState.ToDateRangeInput().Prefix(Icons.CalendarRange), rangeState.ToDateRangeInput().Suffix(Icons.CalendarDays), rangeState.ToDateRangeInput().Prefix(Icons.CalendarRange).Suffix(Icons.CalendarDays))
-               | AffixRow("Bool", boolState.ToBoolInput().Label("Enable").Prefix(Icons.Bell), boolState.ToBoolInput().Label("Enable").Suffix(Icons.Info), boolState.ToBoolInput().Label("Enable").Prefix(Icons.Bell).Suffix(Icons.Info))
-               | AffixRow("Color", colorState.ToColorInput().Prefix(Icons.Palette), colorState.ToColorInput().Suffix(Icons.Pipette), colorState.ToColorInput().Prefix(Icons.Palette).Suffix(Icons.Pipette))
-               | AffixRow("Feedback", feedbackState.ToFeedbackInput().Stars().Prefix(Icons.Star), feedbackState.ToFeedbackInput().Stars().Suffix(Icons.MessageSquare), feedbackState.ToFeedbackInput().Stars().Prefix(Icons.Star).Suffix(Icons.MessageSquare))
-               | AffixRow("Icon", iconState.ToIconInput().Prefix(Icons.Search), iconState.ToIconInput().Suffix(Icons.Sparkles), iconState.ToIconInput().Prefix(Icons.Search).Suffix(Icons.Sparkles))
-               | AffixRow("Code", codeState.ToCodeInput().Prefix(Icons.Code), codeState.ToCodeInput().Suffix(Icons.Braces), codeState.ToCodeInput().Prefix(Icons.Code).Suffix(Icons.Braces));
-    }
-
-    private enum Currency
-    {
-        USD,
-        EUR,
-        GBP,
-    }
-
-    private static GridView AffixHeaderRow() =>
-        Layout.Grid().Columns(4)
-               | null!
-               | Text.Monospaced("Prefix only")
-               | Text.Monospaced("Suffix only")
-               | Text.Monospaced("Both");
-
-    private static GridView AffixRow(string label, object prefixOnly, object suffixOnly, object both) =>
-        Layout.Grid().Columns(4)
-               | Text.Monospaced(label)
-               | prefixOnly
-               | suffixOnly
-               | both;
 }
